+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # Changelog
 
+## v0.2.0 (2024-10-27)
+* Add `lenient` and `conjunction_mode` options to `parse/3`.
+* Add `lenient` to `parse_with_field/4`.
+
 ## v0.1.0 (2024-10-19)
 Initial Release.